requirements.txt
flask
requests
beautifulsoup4
readability-lxml
langchain-core
langchain-google-genai
serverless-wsgi
python-dotenv

